# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserPreferencesPace =
  (typeof UserPreferencesPace)[keyof typeof UserPreferencesPace];

export const UserPreferencesPace = {
  easy: "easy",
  moderate: "moderate",
  active: "active",
} as const;

export type UserPreferencesBudgetLevel =
  (typeof UserPreferencesBudgetLevel)[keyof typeof UserPreferencesBudgetLevel];

export const UserPreferencesBudgetLevel = {
  budget: "budget",
  mid: "mid",
  luxury: "luxury",
} as const;

export interface UserPreferences {
  pace: UserPreferencesPace;
  maxStepsPerDay?: number;
  dietaryNeeds?: string[];
  interests?: string[];
  budgetLevel?: UserPreferencesBudgetLevel;
  accessibilityNeeds?: string[];
}

export interface GenerateItineraryBody {
  city: string;
  country: string;
  /**
   * @minimum 3
   * @maximum 7
   */
  days: number;
  travelMonth: string;
  preferences: UserPreferences;
}

export type CreateItineraryBodyGeneratedData = { [key: string]: unknown };

export interface CreateItineraryBody {
  userId?: string;
  title: string;
  city: string;
  country: string;
  days: number;
  travelMonth: string;
  generatedData: CreateItineraryBodyGeneratedData;
  preferences?: UserPreferences;
}

export type ItineraryGeneratedData = { [key: string]: unknown };

export interface Itinerary {
  id: string;
  userId?: string | null;
  title: string;
  city: string;
  country: string;
  days: number;
  travelMonth: string;
  generatedData: ItineraryGeneratedData;
  createdAt: string;
  updatedAt: string;
}

export interface DayActivity {
  name: string;
  description: string;
  duration: string;
  walkingMinutes: number;
  steps: number;
  cost: string;
  tips?: string;
  isRestStop?: boolean;
}

export interface Restaurant {
  name: string;
  cuisine: string;
  priceRange: string;
  description: string;
  wheelchairFriendly?: boolean;
  nearbyAttraction?: string;
}

export interface TransportOption {
  mode: string;
  description: string;
  estimatedCost: string;
  accessibilityNotes?: string;
}

export interface SideTrip {
  name: string;
  description: string;
  distance: string;
  extraSteps: number;
  extraTime: string;
  estimatedCost: string;
}

export interface DayPlan {
  dayNumber: number;
  theme: string;
  morning: DayActivity[];
  afternoon: DayActivity[];
  evening: DayActivity[];
  totalSteps: number;
  totalWalkingMinutes: number;
  activeHours: number;
  estimatedCostLow: number;
  estimatedCostHigh: number;
  currency: string;
  restaurants: Restaurant[];
  transportOptions: TransportOption[];
  sideTrips: SideTrip[];
  crowdAvoidanceTip?: string;
  weatherNote?: string;
}

export interface GeneratedItinerary {
  title: string;
  city: string;
  country: string;
  days: number;
  travelMonth: string;
  overview: string;
  seniorFriendlyScore: number;
  seniorFriendlyNotes?: string;
  totalEstimatedCostLow: number;
  totalEstimatedCostHigh: number;
  currency: string;
  dayPlans: DayPlan[];
  weatherInfo?: string;
  bestTimeToVisit?: string;
  emergencyNumbers?: string;
}

export interface Destination {
  id: string;
  city: string;
  country: string;
  description: string;
  seniorFriendlyScore: number;
  highlights: string[];
  bestMonths: string[];
  terrain: string;
  imageUrl?: string | null;
}

export type GetItinerariesParams = {
  userId?: string;
};

export type SearchDestinationsParams = {
  query: string;
};