# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List saved itineraries
 */
export const GetItinerariesQueryParams = zod.object({
  userId: zod.coerce.string().optional(),
});

export const GetItinerariesResponseItem = zod.object({
  id: zod.string(),
  userId: zod.string().nullish(),
  title: zod.string(),
  city: zod.string(),
  country: zod.string(),
  days: zod.number(),
  travelMonth: zod.string(),
  generatedData: zod.record(zod.string(), zod.unknown()),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const GetItinerariesResponse = zod.array(GetItinerariesResponseItem);

/**
 * @summary Save an itinerary
 */
export const CreateItineraryBody = zod.object({
  userId: zod.string().optional(),
  title: zod.string(),
  city: zod.string(),
  country: zod.string(),
  days: zod.number(),
  travelMonth: zod.string(),
  generatedData: zod.record(zod.string(), zod.unknown()),
  preferences: zod
    .object({
      pace: zod.enum(["easy", "moderate", "active"]),
      maxStepsPerDay: zod.number().optional(),
      dietaryNeeds: zod.array(zod.string()).optional(),
      interests: zod.array(zod.string()).optional(),
      budgetLevel: zod.enum(["budget", "mid", "luxury"]).optional(),
      accessibilityNeeds: zod.array(zod.string()).optional(),
    })
    .optional(),
});

/**
 * @summary Get itinerary by ID
 */
export const GetItineraryParams = zod.object({
  id: zod.coerce.string(),
});

export const GetItineraryResponse = zod.object({
  id: zod.string(),
  userId: zod.string().nullish(),
  title: zod.string(),
  city: zod.string(),
  country: zod.string(),
  days: zod.number(),
  travelMonth: zod.string(),
  generatedData: zod.record(zod.string(), zod.unknown()),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete an itinerary
 */
export const DeleteItineraryParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Generate a new AI itinerary
 */
export const generateItineraryBodyDaysMin = 3;
export const generateItineraryBodyDaysMax = 7;

export const GenerateItineraryBody = zod.object({
  city: zod.string(),
  country: zod.string(),
  days: zod
    .number()
    .min(generateItineraryBodyDaysMin)
    .max(generateItineraryBodyDaysMax),
  travelMonth: zod.string(),
  preferences: zod.object({
    pace: zod.enum(["easy", "moderate", "active"]),
    maxStepsPerDay: zod.number().optional(),
    dietaryNeeds: zod.array(zod.string()).optional(),
    interests: zod.array(zod.string()).optional(),
    budgetLevel: zod.enum(["budget", "mid", "luxury"]).optional(),
    accessibilityNeeds: zod.array(zod.string()).optional(),
  }),
});

export const GenerateItineraryResponse = zod.object({
  title: zod.string(),
  city: zod.string(),
  country: zod.string(),
  days: zod.number(),
  travelMonth: zod.string(),
  overview: zod.string(),
  seniorFriendlyScore: zod.number(),
  seniorFriendlyNotes: zod.string().optional(),
  totalEstimatedCostLow: zod.number(),
  totalEstimatedCostHigh: zod.number(),
  currency: zod.string(),
  dayPlans: zod.array(
    zod.object({
      dayNumber: zod.number(),
      theme: zod.string(),
      morning: zod.array(
        zod.object({
          name: zod.string(),
          description: zod.string(),
          duration: zod.string(),
          walkingMinutes: zod.number(),
          steps: zod.number(),
          cost: zod.string(),
          tips: zod.string().optional(),
          isRestStop: zod.boolean().optional(),
        }),
      ),
      afternoon: zod.array(
        zod.object({
          name: zod.string(),
          description: zod.string(),
          duration: zod.string(),
          walkingMinutes: zod.number(),
          steps: zod.number(),
          cost: zod.string(),
          tips: zod.string().optional(),
          isRestStop: zod.boolean().optional(),
        }),
      ),
      evening: zod.array(
        zod.object({
          name: zod.string(),
          description: zod.string(),
          duration: zod.string(),
          walkingMinutes: zod.number(),
          steps: zod.number(),
          cost: zod.string(),
          tips: zod.string().optional(),
          isRestStop: zod.boolean().optional(),
        }),
      ),
      totalSteps: zod.number(),
      totalWalkingMinutes: zod.number(),
      activeHours: zod.number(),
      estimatedCostLow: zod.number(),
      estimatedCostHigh: zod.number(),
      currency: zod.string(),
      restaurants: zod.array(
        zod.object({
          name: zod.string(),
          cuisine: zod.string(),
          priceRange: zod.string(),
          description: zod.string(),
          wheelchairFriendly: zod.boolean().optional(),
          nearbyAttraction: zod.string().optional(),
        }),
      ),
      transportOptions: zod.array(
        zod.object({
          mode: zod.string(),
          description: zod.string(),
          estimatedCost: zod.string(),
          accessibilityNotes: zod.string().optional(),
        }),
      ),
      sideTrips: zod.array(
        zod.object({
          name: zod.string(),
          description: zod.string(),
          distance: zod.string(),
          extraSteps: zod.number(),
          extraTime: zod.string(),
          estimatedCost: zod.string(),
        }),
      ),
      crowdAvoidanceTip: zod.string().optional(),
      weatherNote: zod.string().optional(),
    }),
  ),
  weatherInfo: zod.string().optional(),
  bestTimeToVisit: zod.string().optional(),
  emergencyNumbers: zod.string().optional(),
});

/**
 * @summary Get popular destinations
 */
export const GetDestinationsResponseItem = zod.object({
  id: zod.string(),
  city: zod.string(),
  country: zod.string(),
  description: zod.string(),
  seniorFriendlyScore: zod.number(),
  highlights: zod.array(zod.string()),
  bestMonths: zod.array(zod.string()),
  terrain: zod.string(),
  imageUrl: zod.string().nullish(),
});
export const GetDestinationsResponse = zod.array(GetDestinationsResponseItem);

/**
 * @summary Search destinations
 */
export const SearchDestinationsQueryParams = zod.object({
  query: zod.coerce.string(),
});

export const SearchDestinationsResponseItem = zod.object({
  id: zod.string(),
  city: zod.string(),
  country: zod.string(),
  description: zod.string(),
  seniorFriendlyScore: zod.number(),
  highlights: zod.array(zod.string()),
  bestMonths: zod.array(zod.string()),
  terrain: zod.string(),
  imageUrl: zod.string().nullish(),
});
export const SearchDestinationsResponse = zod.array(
  SearchDestinationsResponseItem,
);