# lib/api-zod/src/generated/types/createItineraryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateItineraryBodyGeneratedData } from "./createItineraryBodyGeneratedData";
import type { UserPreferences } from "./userPreferences";

export interface CreateItineraryBody {
  userId?: string;
  title: string;
  city: string;
  country: string;
  days: number;
  travelMonth: string;
  generatedData: CreateItineraryBodyGeneratedData;
  preferences?: UserPreferences;
}