# lib/api-zod/src/generated/types/dayActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DayActivity {
  name: string;
  description: string;
  duration: string;
  walkingMinutes: number;
  steps: number;
  cost: string;
  tips?: string;
  isRestStop?: boolean;
}