# lib/api-zod/src/generated/types/dayPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DayActivity } from "./dayActivity";
import type { Restaurant } from "./restaurant";
import type { SideTrip } from "./sideTrip";
import type { TransportOption } from "./transportOption";

export interface DayPlan {
  dayNumber: number;
  theme: string;
  morning: DayActivity[];
  afternoon: DayActivity[];
  evening: DayActivity[];
  totalSteps: number;
  totalWalkingMinutes: number;
  activeHours: number;
  estimatedCostLow: number;
  estimatedCostHigh: number;
  currency: string;
  restaurants: Restaurant[];
  transportOptions: TransportOption[];
  sideTrips: SideTrip[];
  crowdAvoidanceTip?: string;
  weatherNote?: string;
}