# lib/api-zod/src/generated/types/destination.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Destination {
  id: string;
  city: string;
  country: string;
  description: string;
  seniorFriendlyScore: number;
  highlights: string[];
  bestMonths: string[];
  terrain: string;
  imageUrl?: string | null;
}