# lib/api-zod/src/generated/types/generateItineraryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPreferences } from "./userPreferences";

export interface GenerateItineraryBody {
  city: string;
  country: string;
  /**
   * @minimum 3
   * @maximum 7
   */
  days: number;
  travelMonth: string;
  preferences: UserPreferences;
}