# lib/api-zod/src/generated/types/generatedItinerary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DayPlan } from "./dayPlan";

export interface GeneratedItinerary {
  title: string;
  city: string;
  country: string;
  days: number;
  travelMonth: string;
  overview: string;
  seniorFriendlyScore: number;
  seniorFriendlyNotes?: string;
  totalEstimatedCostLow: number;
  totalEstimatedCostHigh: number;
  currency: string;
  dayPlans: DayPlan[];
  weatherInfo?: string;
  bestTimeToVisit?: string;
  emergencyNumbers?: string;
}