# lib/api-zod/src/generated/types/itinerary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryGeneratedData } from "./itineraryGeneratedData";

export interface Itinerary {
  id: string;
  userId?: string | null;
  title: string;
  city: string;
  country: string;
  days: number;
  travelMonth: string;
  generatedData: ItineraryGeneratedData;
  createdAt: Date;
  updatedAt: Date;
}