# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Restaurant {
  name: string;
  cuisine: string;
  priceRange: string;
  description: string;
  wheelchairFriendly?: boolean;
  nearbyAttraction?: string;
}