# lib/api-zod/src/generated/types/sideTrip.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SideTrip {
  name: string;
  description: string;
  distance: string;
  extraSteps: number;
  extraTime: string;
  estimatedCost: string;
}