# lib/api-zod/src/generated/types/transportOption.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TransportOption {
  mode: string;
  description: string;
  estimatedCost: string;
  accessibilityNotes?: string;
}