# lib/api-zod/src/generated/types/userPreferences.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorTravel API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPreferencesBudgetLevel } from "./userPreferencesBudgetLevel";
import type { UserPreferencesPace } from "./userPreferencesPace";

export interface UserPreferences {
  pace: UserPreferencesPace;
  maxStepsPerDay?: number;
  dietaryNeeds?: string[];
  interests?: string[];
  budgetLevel?: UserPreferencesBudgetLevel;
  accessibilityNeeds?: string[];
}